+++ Resources/Prototypes/Entities/Objects/Misc/space_cash.yml
@@ -34,9 +34,9 @@
     - cash_50000 # Frontier: larger denominations
     - cash_100000 # Frontier: larger denominations
     - cash_250000 # Frontier: larger denominations (cash_1000000<cash_250000)
-    layerFunction: # Frontier: multicolour cash
-      !type:ThresholdStackLayerFunction # Frontier
-        thresholds: [10, 100, 500, 1000, 5000, 10000, 25000, 50000, 100000, 250000] # Frontier
+    layerFunction: Threshold # Frontier: multicolour cash
+  - type: ThresholdStackLayerFunction # Frontier
+    thresholds: [10, 100, 500, 1000, 5000, 10000, 25000, 50000, 100000, 250000] # Frontier
   - type: Sprite
     sprite: _NF/Objects/Economy/cash.rsi # Frontier: larger denominations
     state: cash
